requirements.txt
-i https://pypi.org/simple
asgiref==3.6.0 ; python_version >= '3.7'
contourpy==1.0.7 ; python_version >= '3.8'
cycler==0.11.0 ; python_version >= '3.6'
django==3.2
django-filter==23.1
djangorestframework==3.14.0
djongo==1.3.6
fonttools==4.39.3 ; python_version >= '3.8'
gunicorn==20.1.0
importlib-resources==5.12.0 ; python_version < '3.10'
joblib==1.2.0
kiwisolver==1.4.4 ; python_version >= '3.7'
matplotlib==3.7.1 ; python_version >= '3.8'
numpy==1.24.2
packaging==23.1 ; python_version >= '3.7'
pandas==2.0.0
pillow==9.5.0 ; python_version >= '3.7'
pymongo==3.12.1
pyparsing==3.0.9 ; python_full_version >= '3.6.8'
python-dateutil==2.8.2 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pytz==2023.3
scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.12.2
setuptools==67.6.1 ; python_version >= '3.7'
six==1.16.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sqlparse==0.2.4 ; python_version >= '3.5'
threadpoolctl==3.1.0 ; python_version >= '3.6'
tzdata==2023.3 ; python_version >= '2'
zipp==3.15.0 ; python_version < '3.10'
